Exported basis vector widget
